Update README.md
BitComics is an alternative to create exponential value to sequential arts, literary and other creative narratives, in addition to a geek community based on blockchain. Above all, bitcomics is an initiative dedicated to comics, blockchain, NFTs and the Geek world.
Non-fungible tokens (NFTs) make it technically possible for digital assets to be owned and traded, introducing the concept of scarcity into the digital realm. An NFT can be transferred anywhere in the world in minutes, which presents a huge advantage, guaranteeing an official and unique registration of a certain asset.
It is important to emphasize that comics are becoming increasingly targeted as the result of a creative and profitable product. Comic books are part of the so-called Creative Economy, the name given to a set of activities and services that originate in intellectual capital.
With the demand of consumers increasingly avid for exclusivity, we have endless opportunities. Therefore, a decentralized publisher through blockchain technology has the ability to solve this real problem. The construction of a decentralized publisher will provide the possibility for publishing, selling and buying comics, where authors not only hold the rights to their characters and works, but also the power to sell them at the desired price, choosing the number of copies to be sold in the form of NFTs, or even services linked to their arts, in addition to royalties for their reproduction or negotiations. Easy to operate, our platform will allow publishing at low rates 100% on demand. That is, the user determines how much he wants to earn per sale by making it available virtually.